+++ Editor/Content/GeometryMetadata.cs
@@ -2,6 +2,7 @@
 {
     public class GeometryMetadata : AssetMetadata
     {
+        public string Name { get; init; }
         public List<LodInfo> LODs { get; init; }
     }
 }